backed testing done successful minor issue are in log_management and nlp_query will check weather its llm issue or something else

+++ src/app/logout/page.tsx
@@ -1,13 +1,21 @@
 "use client"
 import { Button } from "@/components/ui/button"
 import { useRouter } from "next/navigation"
+import axios from "axios"
 export default function Logout() {
 
     const router = useRouter()
 
     const handleLogout = async () => {
-        localStorage.clear()
-        router.push('/llm_config')}
+        const res = await axios.get('/api/users/logout')
+        if (res.data.success) {
+            localStorage.clear();
+            router.push('/llm_config')
+        }
+        else{
+            alert("Logout failed")
+        }
+    }
     return (
         <div className="min-h-screen flex items-center justify-center bg-gray-100 px-4 py-12 sm:px-6 lg:px-8">
         <Button onClick={handleLogout}>Logout</Button>
